pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "archivist"
version = "0.1.0"
description = "A self-hosted corpus-building tool for semantic search and RAG"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "1507 Systems" },
]
keywords = ["corpus", "embeddings", "semantic-search", "rag", "mcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: Indexing",
]

dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "pymupdf>=1.24",
    "feedparser>=6.0",
    "sentence-transformers>=3.0",
    "chromadb>=0.5",
    "mcp>=1.0",
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "rich>=13.0",
]

[project.optional-dependencies]
whisper = ["openai-whisper>=20231117"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "mypy>=1.10",
    "ruff>=0.5",
    "types-PyYAML",
    "types-beautifulsoup4",
    "respx>=0.21",
]

[project.scripts]
archivist = "archivist.cli:cli"

[project.urls]
Homepage = "https://github.com/1507-systems/archivist"
Repository = "https://github.com/1507-systems/archivist"
Issues = "https://github.com/1507-systems/archivist/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"
